étape 10 start

+++ src/ch/epfl/rigel/gui/ViewingParametersBean.java
@@ -1,28 +1,24 @@
 package ch.epfl.rigel.gui;
 
 import ch.epfl.rigel.coordinates.HorizontalCoordinates;
-import javafx.beans.property.DoubleProperty;
-import javafx.beans.property.ObjectProperty;
-import javafx.beans.property.SimpleDoubleProperty;
-import javafx.beans.property.SimpleObjectProperty;
+import javafx.beans.property.*;
 
-import java.time.LocalDate;
 
 
 public final class ViewingParametersBean {
 
-    private final DoubleProperty fieldOfViewDeg = new SimpleDoubleProperty();
+    private final IntegerProperty fieldOfViewDeg = new SimpleIntegerProperty();
     private final ObjectProperty<HorizontalCoordinates> center = new SimpleObjectProperty<>(null);
 
-    public DoubleProperty fieldOfViewDegProperty(){
+    public IntegerProperty fieldOfViewDegProperty(){
         return fieldOfViewDeg;
     }
 
-    public Double getFieldOfViewDeg(){
+    public int getFieldOfViewDeg(){
         return fieldOfViewDeg.get();
     }
 
-    public void setFieldOfViewDeg(Double fieldOfViewDeg){
+    public void setFieldOfViewDeg(int fieldOfViewDeg){
         this.fieldOfViewDeg.set(fieldOfViewDeg);
     }
